[LEMONEY-14]
- Adding a confirmation dialog to the delete button on offers datagrid.

+++ app/javascript/components/admin/offers/delete_button/delete_button.jsx
@@ -12,6 +12,10 @@ class DeleteButton extends React.Component {
 	handleClick = (e) => {
 		e.preventDefault();
 
+		const actionConfirmed = confirm('This offer will be deleted. Do you wish to proceed?');
+
+		if (!actionConfirmed) return;
+
 		const requestUrl = `${Api.address}/offers/${this.props.resourceId}`;
 		const requestOptions = {method: 'DELETE'};
